+++ PresentataionSwiftUITests/PresentataionSwiftUITests.swift
@@ -1,0 +1,17 @@
+//
+//  PresentataionSwiftUITests.swift
+//  PresentataionSwiftUITests
+//
+//  Created by 조영태 on 7/25/25.
+//
+
+import Testing
+@testable import PresentataionSwiftUI
+
+struct PresentataionSwiftUITests {
+
+    @Test func example() async throws {
+        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+    }
+
+}
