+++ ko.lproj/Localizable.strings
@@ -21,3 +21,4 @@ retry = "재시도";
 todoNotFound = "할일을 찾을 수 없습니다";
 unKnownError = "알 수 없는 에러입니다";
 error = "오류";
+
