+++ todoApp/DI/Container/SwiftUI/TodoListDIContainer.SUI.swift
@@ -10,9 +10,10 @@ import Swinject
 import DataLayer
 import Domain
 import PresentationShared
+import PresentataionSwiftUI
 
 extension SUI {
-    class TodoListDIContainer {
+    class TodoListDIContainer: TodoListDIContainerProcotcol{
         private let container: Container
         private let repoAssembly: TodoRepositoryAssembly
         private let useCaseAssembly: TodoUseCaseAssembly
@@ -36,8 +37,8 @@ extension SUI {
         func makeTodoListScene(
             initFilter: TodoFilterType,
             env: DataEnvironment = .local,
-            setup: ((_ vm: SUI.TodoListVM) -> Void)? = nil
-        ) -> SUI.TodoListVC {
+            setup: ((_ vm: TodoListVM) -> Void)? = nil
+        ) -> TodoListVC {
             let vm = makeTodoListVM(initFilter: initFilter, env: env)
             setup?(vm)
             return makeTodoListVC(vm)
@@ -45,43 +46,31 @@ extension SUI {
 
         private func makeTodoListVM(
             initFilter: TodoFilterType, env: DataEnvironment = .local
-        ) -> SUI.TodoListVM {
-            let repo = container.resolveOrFail(
-                TodoRepository.self,
-                argument: env
-            )
+        ) -> TodoListVM {
+            let ds = repoAssembly.makeDataSource(.local)
+            let repo = repoAssembly.makeRepository(ds)
 
-            let cache = container.resolveOrFail(TodoListCacheUseCase.self)
-            let fetchUseCase = container.resolveOrFail(
-                (any FetchTodoUseCase).self, argument: repo
-            )
-            let deleteUseCase = container.resolveOrFail(
-                (any DeleteTodoUseCase).self,
-                arguments: repo, cache
-            )
-            let toggleDone = container.resolveOrFail(
-                (any ToggleTodoDoneUseCase).self,
-                arguments: repo, cache
-            )
-
-            let useCase = SUI.TodoListVM.UseCase(
+            let cache = useCaseAssembly.makeListCacheUseCase()
+            
+            let fetchUseCase = useCaseAssembly.makeFetchTodoUseCase(repo)
+            let deleteUseCase = useCaseAssembly.makeDeleteTodoUseCase(repo, cache)
+            let toggleDone = useCaseAssembly.makeToggleTodoDoneUseCase(repo, cache)
+            let useCase = TodoListVM.UseCase(
                 fetch: fetchUseCase,
                 delete: deleteUseCase,
                 toggleDone: toggleDone,
                 cache: cache
             )
 
-            return self.container
-                .resolveOrFail(
-                    SUI.TodoListVM.self,
-                    arguments: initFilter,
-                    useCase
-                )
+            return self.container.resolveOrFail(
+                TodoListVM.self,
+                arguments: initFilter,useCase
+            )
         }
 
-        private func makeTodoListVC(_ vm: SUI.TodoListVM) -> SUI.TodoListVC {
+        private func makeTodoListVC(_ vm: TodoListVM) -> TodoListVC {
             return self.container.resolveOrFail(
-                SUI.TodoListVC.self,
+                TodoListVC.self,
                 argument: vm
             )
         }
@@ -90,23 +79,23 @@ extension SUI {
     final private class TodoListAssembly: Assembly {
         func assemble(container: Container) {
             // swiftUI vm 등록
-            container.register(SUI.TodoListVM.self) {
+            container.register(TodoListVM.self) {
                 (
                     resolver,
                     initFilter: TodoFilterType,
-                    useCase: SUI.TodoListVM.UseCase
+                    useCase: TodoListVM.UseCase
                 ) in
 
-                return SUI.TodoListVM(useCase, initFilter: initFilter)
+                return TodoListVM(useCase, initFilter: initFilter)
             }
             // swiftUI vc 등록
-            container.register(SUI.TodoListVC.self) {
+            container.register(TodoListVC.self) {
                 (
                     resolver,
-                    viewModel: SUI.TodoListVM
+                    viewModel: TodoListVM
                 ) in
                     
-                return SUI.TodoListVC(viewModel: viewModel)
+                return TodoListVC(viewModel: viewModel)
             }
         }
     }
