+++ todoApp/DI/Container/UIKit/AppDIContainer.UIK.swift
@@ -7,28 +7,25 @@
 
 import Foundation
 import Swinject
-import PresentationUIKit
 import UIKit
+import PresentationUIKit
 
 extension UIK {
-    final class AppDIContainer: AppDIContainerProtocol {
-        static let shared = AppDIContainer()
-
-        private let container = Container()
-
-        init() {
-            _ = Assembler(
-                [
-                    // 전역 공통 의존성 (예: 네트워크, 유틸 등)
-                ],
-                container: container
-            )
-        }
+    final class AppDIContainer: DefaultAppDIContainer, AppDIContainerProtocol {
+        static let _shared:AppDIContainer? = nil
         
+        static var shared: AppDIContainer {
+            if let shared = _shared {
+                return shared
+            }
+            
+            return AppDIContainer()
+        }
+
         func makeTodoListDIContainer() -> any TodoListDIContainerProtocol {
             return UIK.TodoListDIContainer(parentContainer: container)
         }
-        
+    
         func makeWritableDIContainer() -> any WritableTodoDIContainerProtocol {
             return UIK.WritableTodoDIContainer(parentContainer: container)
         }
@@ -47,64 +44,5 @@ extension UIK {
             
             coordinator.start()
         }
-    }
-}
-
-extension Resolver {
-    func resolveOrFail<Service>(
-        _ serviceType: Service.Type, name: String? = nil
-    ) -> Service {
-        guard let resolved = self.resolve(serviceType, name: name) else {
-            preconditionFailure(
-                #function + "❌ Failed to resolve: \(serviceType)")
-        }
-        return resolved
-    }
-
-    func resolveOrFail<Service, Arg>(
-        _ serviceType: Service.Type, argument: Arg, name: String? = nil
-    ) -> Service {
-        guard
-            let resolved = self.resolve(
-                serviceType, name: name, argument: argument)
-        else {
-            preconditionFailure(
-                #function
-                    + "❌ Failed to resolve: \(serviceType) with argument: \(Arg.self)"
-            )
-        }
-        return resolved
-    }
-
-    func resolveOrFail<Service, Arg1, Arg2>(
-        _ serviceType: Service.Type, arguments arg1: Arg1, _ arg2: Arg2,
-        name: String? = nil
-    ) -> Service {
-        guard
-            let resolved = self.resolve(
-                serviceType, name: name, arguments: arg1, arg2)
-        else {
-            preconditionFailure(
-                #function
-                    + "❌ Failed to resolve: \(serviceType) with arguments: \(Arg1.self), \(Arg2.self)"
-            )
-        }
-        return resolved
-    }
-    
-    func resolveOrFail<Service, Arg1, Arg2, Arg3>(
-        _ serviceType: Service.Type, arguments arg1: Arg1, _ arg2: Arg2, _ arg3: Arg3,
-        name: String? = nil
-    ) -> Service {
-        guard
-            let resolved = self.resolve(
-                serviceType, name: name, arguments: arg1, arg2, arg3)
-        else {
-            preconditionFailure(
-                #function
-                    + "❌ Failed to resolve: \(serviceType) with arguments: \(Arg1.self), \(Arg2.self), \(Arg3.self)"
-            )
-        }
-        return resolved
     }
 }
