--- todoApp/Protocol.swift
@@ -1,21 +1,0 @@
-////
-////  IOProtocol.swift
-////  todoApp
-////
-////  Created by 조영태 on 5/1/25.
-////
-//
-//import Foundation
-//import RxSwift
-//
-//struct IOEmpty {}
-//
-//protocol HasRxIO {
-//    associatedtype Input
-//    associatedtype Output
-//
-////    var input: Input { get }
-////    var output: Output { get }
-//    
-//    var disposeBag: DisposeBag { get }
-//}
